requirements-to-freeze.txt
boto3==1.4.4
botocore==1.5.60
click==6.6
docutils==0.12
ExifRead==2.1.2
jmespath==0.9.0
pyexifinfo==0.4.0
python-dateutil==2.5.3
python-dotenv==0.6.0
python-magic==0.4.12
redis==2.10.5
s3transfer==0.1.10
six==1.10.0
